рефактор кода + новая версия 0.5
