+++ gradle.properties
@@ -9,7 +9,7 @@ yarn_mappings=1.20.1+build.10
 loader_version=0.15.11
 
 # Mod Properties
-mod_version=0.4.2-1.20.1
+mod_version=0.5-1.20.1
 maven_group=com.dota2
 archives_base_name=dotacraft
