+++ src/main/java/com/dota2/event/ModEvents.java
@@ -1,30 +1,12 @@
 package com.dota2.event;
 
-import com.dota2.mixin.PlayerEntityMixin;
 import net.fabricmc.fabric.api.entity.event.v1.ServerLivingEntityEvents;
-import net.minecraft.entity.player.PlayerEntity;
-
-import static com.dota2.component.ModComponents.HERO_COMPONENT;
+import net.fabricmc.fabric.api.event.player.AttackEntityCallback;
 
 public class ModEvents {
 
-    public static void AllowDamage() {
-        ServerLivingEntityEvents.ALLOW_DAMAGE.register(((entity, source, amount) -> {
-
-            if (entity.isPlayer()) {
-                PlayerEntity player = (PlayerEntity) entity;
-                if (player.getComponent(HERO_COMPONENT).isHero()) {
-                    player.setHealth(20);
-                    player.setFireTicks(0);
-                    player.setFrozenTicks(0);
-                    return false;
-                } else {
-                    return true;
-                }
-            } else {
-                return true;
-            }
-        }));
-
+    public static void registerEvents() {
+        ServerLivingEntityEvents.ALLOW_DAMAGE.register(AllowDamage::allowDamage);
+        AttackEntityCallback.EVENT.register(AttackEntity::onAttackEntity);
     }
 }
