+++ src/main/java/com/dota2/util/InventoryUtils.java
@@ -1,5 +1,7 @@
 package com.dota2.util;
 
+import com.dota2.item.Weapon;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
@@ -7,13 +9,27 @@ import net.minecraft.item.ItemStack;
 public class InventoryUtils {
 
     public static int getFirstInventoryIndex(PlayerInventory inventory, Item item) {
-        for(int i = 0; i < inventory.size(); i++) {
+        for (int i = 0; i < inventory.size(); i++) {
             ItemStack currentStack = inventory.getStack(i);
             if (!currentStack.isEmpty() && currentStack.isOf(item)) {
                 return i;
             }
         }
         return -1;
+    }
+
+    public static int calculateDamage(PlayerEntity provider) {
+        int totalDamage = 0;
+
+        for (int i = 0; i < provider.getInventory().size(); i++) {
+            ItemStack stack = provider.getInventory().getStack(i);
+
+            if (stack.getItem() instanceof Weapon weapon) {
+                totalDamage += weapon.getDamage();
+            }
+        }
+
+        return totalDamage;
     }
 
 }
