+++ src/main/resources/assets/dotacraft/lang/en_us.json
@@ -18,11 +18,11 @@
   "itemgroup.dota": "Dota Craft",
   "block.dotacraft.portal": "Portal",
   "block.dotacraft.pressure_plate": "Pressure Plate",
-  "effect.dotacraft.bottle_regeneration_mana":  "Regeneration of mana",
+  "effect.dotacraft.bottle_regeneration_mana": "Regeneration of mana",
   "effect.dotacraft.clarity_regeneration_mana": "Regeneration of mana",
   "effect.dotacraft.bottle_regeneration_health": "Regeneration of health",
-  "effect.dotacraft.flask_regeneration_health":  "Regeneration of health",
-  "effect.dotacraft.tango_regeneration_health":  "Regeneration of health",
+  "effect.dotacraft.flask_regeneration_health": "Regeneration of health",
+  "effect.dotacraft.tango_regeneration_health": "Regeneration of health",
   "attribute.name.generic.crit_chance": "Chance of critical damage",
   "text.dotacraft.tango_timer": "%d left"
 }
