+++ src/main/resources/assets/dotacraft/lang/ru_ru.json
@@ -18,11 +18,11 @@
   "itemgroup.dota": "Dota Craft",
   "block.dotacraft.portal": "Портал",
   "block.dotacraft.pressure_plate": "Нажимная плита",
-  "effect.dotacraft.bottle_regeneration_mana":  "Регенерация маны",
+  "effect.dotacraft.bottle_regeneration_mana": "Регенерация маны",
   "effect.dotacraft.clarity_regeneration_mana": "Регенерация маны",
   "effect.dotacraft.bottle_regeneration_health": "Регенерация здоровья",
-  "effect.dotacraft.flask_regeneration_health":  "Регенерация здоровья",
-  "effect.dotacraft.tango_regeneration_health":  "Регенерация здоровья",
+  "effect.dotacraft.flask_regeneration_health": "Регенерация здоровья",
+  "effect.dotacraft.tango_regeneration_health": "Регенерация здоровья",
   "attribute.name.generic.crit_chance": "Шанс критического удара",
   "text.dotacraft.tango_timer": "Осталось %d"
 }
